check_megacli_diskgroups: Now check present disk groups with number of expected ones
The number of expected disk groups can be set by the variable
`num_disk_groups` in /etc/check_mk/megacli.cfg (which again is set by
the role variable `checkmkagent_megacli_num_disk_groups`.

The number of expected and present disk groups was not checked before,
now we have a CRIT state if the numbers do not match.