checkmk-plugins.yml: Add checkmkagent_mk_apt_update + checkmkagent_mk_apt_delay
The agent plugin mk_apt can be configured to run `apt update` before checking for apt updates.
(Note, that this is not needed nowadays as there exists a systemd timer apt-daily.timer
which performs `apt updates` twice a day.)

The make the agent plugin mk_apt async, a delay has to be set.
This is usually only needed when 'checkmkagent_mk_apt_update' is set to 'yes'.

The workaround for the get_url bug ansible/ansible#65687 is not needed
for the deployment of "mk_apt" as we have to modify the file localhost
before uploading. Only if the file differs a change should be shown.

Also set the default checkmkagent_version to the current latest
(old)stable version.