Improve workaround for get_url bug ansible/ansible#65687
The workaround for the get_url bug ansible/ansible#65687 is still
needed, but the destination file didn't get updated.

If `force` is yes, and dest is not a directory, the file is downloaded
every time and replaces the file if the contents change (`force` defaults
to no).

If `dest` is a directory, the file will always be downloaded (regardless
of the force option), but replaced only if the contents changed.

So we have two options: either set `force` to yes or use a directory as
destination. As the the filename is not changed, decided to only use the
directory as `dest`.

Bumped the default checkmkagent_version to 2.0.0p30, our current latest
checkmk version in use.