Ensure check-mk-agent port 6556 is closed
xinetd needs some time to restart.
When cmk-agent-ctl-daemon.service gets started a race condition is
possible where the check-mk-agent port 6556 has not been freed because
xinet has not restarted completely.

To prevent that, we make sure that port 6556 is closed before we start
cmk-agent-ctl-daemon.service (but only if check-mk-agent was running via
xinetd).