Check if xinetd is installed
xinetd cannot be restarted if it is not installed, but we will try to
restart it when /etc/xinetd.d/check_mk is removed.

To prevent the restart if xinetd is not installed, we register a
xinetd_check variable which is set to "succeeded" if the package is
installed. The variable needs to be registered even in check mode.

The 'xinetd restart' handler will then only be triggered if xinetd is
installed.

Finally, we make sure that the check-mk-agent port 6556 is closed only
if xinetd has been restarted (and thus check-mk-agent has been
disabled).