tasks/checkmkagent-systemd.yml: Rename register variable
We check if the checkmk agent is running in legacy mode, so we renamed
the variable name to make it clearer.

If the checkmk agent has not been upgraded from v2.0.0, the legacy
services will not exist and stopping and disabling the services will
fail. We have to ignore this. If stopping the services causes an
error, we have to handle that in a separate issue.