Enable systemd for podman
podman supports systemd out of the box, but `override_command: false`
(or `command`) needs to be set to make it work, see:
https://github.com/ansible-community/molecule-plugins/blob/main/src/molecule_plugins/podman/driver.py#L108

We can now install the checkmk agent package and without the need to
skip the systemd tasks.

check-mk-agent_2.2.0p9-1_all.deb is about 4MB and needs to be excluded
from the check-added-large-files pre-commit hook.

The ansible default scenario now symlinks to checkmk-agent-v2.2 as a
preparation for multiple scenarios.