Do not change checkmk agent service states in check run
When performing a check run and there is no older version of the checkmk
agent, changing the service state of the checkmk agent fails,
because the package hasn't been actually been installed.

To work around this problem, we do not change the service state when we
perform a check run.