Add support for checkmk agent plugin smart
The CheckMK agent plugins are all listed in
  CheckMK > Setup > Agents > Linux > Plugins

These plugins just can be retrieved via:
- {{ checkmkagent_host_url }}/check_mk/agents/plugins/{{ item }}
- /opt/omd/versions/{{ checkmkagent_version }}.cre/share/check_mk/agents/plugins/{{ item }}

A checkmk agent plugin needs to to the `checkmkagent_plugins_path`
(`/usr/lib/check_mk_agent/plugins`) of the given host.

The basic functionality is already handled in
`tasks/checkmk-deploy-plugins.yml`.

So, we just need to add the plugin to the variable
`checkmkagent_plugins_available` and add the plugin to the hosts
variable `checkmkagent_plugins_enabled` to deploy the agent plugin.